Fix molecule model deletion

+++ src/model/base_model.hpp
@@ -18,12 +18,12 @@ namespace VTX
 		class BaseModel : public Generic::HasCollection<View::BaseView<BaseModel>>, public Selection::BaseSelectable
 		{
 		  public:
-			inline static uint COUNTER = 0;
+			inline static long COUNTER = 0;
 
 			virtual ~BaseModel() = default;
 
-			uint getId() const { return _id; }
-			void setId( const uint p_id ) { _id = p_id; }
+			long getId() const { return _id; }
+			void setId( const long p_id ) { _id = p_id; }
 
 		  protected:
 			virtual void _notifyViews( const Event::VTX_EVENT_MODEL p_event ) final
@@ -34,7 +34,7 @@ namespace VTX
 				}
 			}
 
-			uint _id = COUNTER++;
+			long _id = COUNTER++;
 		};
 	} // namespace Model
 } // namespace VTX
